current-time-insert-py/requirements.txt
pymongo
python-dotenv
pytz
ibm-cos-sdk